% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(mu, sigma, rho, initial, N)
}
\arguments{
\item{mu}{parameter of bivariate normal distribution}

\item{sigma}{parameter of bivariate normal distribution}

\item{rho}{parameter of bivariate normal distribution}

\item{initial}{initial value}

\item{N}{length of chain}
}
\value{
a random sample of size \code{N}
}
\description{
A Gibbs sampler using Rcpp
}
