% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22035R.R
\name{HomeworkPackage}
\alias{HomeworkPackage}
\title{Import packages for homework review file.}
\description{
Import packages for homework review file
}
\examples{
\dontrun{
library(Rcpp)
library(microbenchmark)
library(knitr)
library(rmarkdown)
library(boot)
library(bootstrap)
library(DAAG) 
library(MASS)
library(ggplot2)
}
}
