% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22035R.R
\name{Hybrid.PSO}
\alias{Hybrid.PSO}
\title{Hybrid Particle Swarm Optimization Algorithm}
\usage{
Hybrid.PSO(solveFun, particleNum, limitX, w, c1, c2, iters, alpha)
}
\arguments{
\item{solveFun}{Target function}

\item{particleNum}{Number of particles}

\item{limitX}{The limit of the independent variable}

\item{w}{Weight of inertia}

\item{c1}{Constant of acceleration}

\item{c2}{Constant of acceleration}

\item{iters}{Maximum number of iterations}

\item{alpha}{Value increment threshold of the optimal fitness value}
}
\value{
The optimal (maximum) solution to the target function
}
\description{
Improved PSO Algorithm to handle complex optimization problem
}
\examples{
\dontrun{
solveFun <- function(x){
  log(sin(pi*x)+2)
}
limitX <- c(-1, 2) 
particleNum <- 20 
w <- 1          
c1 <- c2 <- 2   
iters <- 10000  
alpha <- 0.0002
solution <- Hybrid.PSO(solveFun,particleNum,limitX,w,c1,c2,iters,alpha)
}
}
