% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22035R.R
\name{MA}
\alias{MA}
\title{Example data of an adjacent matrix}
\description{
Example data used to illustrate the performance of \code{PageRank}.
}
\examples{
\dontrun{
data(MA)
attach(MA)
NumNode <- 50
nrow <- ncol <- NumNode
IniVec <- runif(NumNode, 0, 1)
Score <- PageRank(MA, 0.85, 100000, IniVec)
}
}
