% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PageRank}
\alias{PageRank}
\title{PageRank Algorithm}
\usage{
PageRank(MA, d, MaxIter, IniVec)
}
\arguments{
\item{MA}{Matrix of adjacency}

\item{d}{damping rate}

\item{MaxIter}{Maximum number of iteration}

\item{IniVec}{Initial iteration vector}
}
\value{
Score result of each node
}
\description{
PageRank Algorithm using Rcpp
}
\examples{
\dontrun{
NumNode <- 50
nrow <- ncol <- NumNode
data(MA)
IniVec <- runif(NumNode, 0, 1)
Score <- PageRank(MA, 0.85, 100000, IniVec)
data <- data.frame(
ndex = 1:NumNode,  
value = Score
)
ggplot(data, aes(x=index, y=value)) + 
geom_bar(stat = "identity", width=0.5)
}
}
