\name{StatComp22035-package}
\alias{StatComp22035-package}
\docType{package}
\title{
  This package provides a homework review and two new functions.
}
\description{
A review of all the homework in course Statistic Computation and two new funtion, 'Hybrid.PSO' and 'PageRank' are introduced in this package.
}
\author{
Hao Yin
Maintainer: Hao Yin <yinhnavi@mail.ustc.edu.cn>
}
\keyword{ package }